requirements.txt
moviepy
chromadb
torch
torchvision
transformers
datasets
flask
flask-cors
llama-cpp-python # cpu only